#168 added media additions module, Image.bitmap moved to this module
