+++ settings.gradle.kts
@@ -25,6 +25,7 @@ if (!pluginPublish) {
         include(":widgets-collection")
         include(":widgets-datetime-picker")
         include(":widgets-image-network")
+        include(":widgets-media")
 
         if (!additionsPublish) {
             include(":sample:android-app")
