+++ widgets-media/build.gradle.kts
@@ -1,0 +1,20 @@
+/*
+ * Copyright 2020 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+plugins {
+    id("com.android.library")
+    id("org.jetbrains.kotlin.multiplatform")
+    id("dev.icerock.mobile.multiplatform")
+    id("maven-publish")
+}
+
+dependencies {
+    mppLibrary(Deps.Libs.MultiPlatform.kotlinStdLib)
+    mppLibrary(Deps.Libs.MultiPlatform.coroutines)
+
+    mppLibrary(Deps.Libs.MultiPlatform.mokoWidgets)
+    mppLibrary(Deps.Libs.MultiPlatform.mokoMedia)
+
+    androidLibrary(Deps.Libs.Android.lifecycle)
+}
