+++ widgets-media/src/iosArm64Main
@@ -1,0 +1,1 @@
+iosMain
