+++ widgets-media/src/iosX64Main
@@ -1,0 +1,1 @@
+iosMain
