+++ widgets/src/commonMain/kotlin/dev/icerock/moko/widgets/core/Image.kt
@@ -4,7 +4,6 @@
 
 package dev.icerock.moko.widgets.core
 
-import dev.icerock.moko.media.Bitmap
 import dev.icerock.moko.resources.ImageResource
 
 expect abstract class Image {
@@ -12,4 +11,3 @@ expect abstract class Image {
 }
 
 expect fun Image.Companion.resource(imageResource: ImageResource): Image
-expect fun Image.Companion.bitmap(bitmap: Bitmap): Image
